Update for Rotation to Orientation rename
