+++ testplugins/src/main/java/org/spongepowered/test/datatest/DataTest.java
@@ -1091,7 +1091,7 @@ public final class DataTest  {
 
         this.checkOfferData(ravager, Keys.ROARING_TIME, Ticks.of(20));
 
-//        this.checkOfferData(itemFrame, Keys.ROTATION, Rotations.LEFT.get());
+//        this.checkOfferData(itemFrame, Keys.ROTATION, Orientations.LEFT.get());
 
         this.checkOfferData(player, Keys.SATURATION, 20.0);
